Feat: Добавляет приватный маршрут для favorites
