+++ src/components/header/header.tsx
@@ -1,4 +1,5 @@
-import {Page} from '../../utils/const.ts';
+import {Link} from 'react-router-dom';
+import {AppRoute, Page} from '../../utils/const.ts';
 import NavBlock from '../nav-block/nav-block.tsx';
 
 type HeaderProps = {
@@ -12,9 +13,9 @@ export default function Header({currentPage, favoritesQuantity}: HeaderProps): J
       <div className="container">
         <div className="header__wrapper">
           <div className="header__left">
-            <a className="header__logo-link header__logo-link--active">
+            <Link className={`header__logo-link ${currentPage === Page.Main ? 'header__logo-link--active' : ''}`} to={AppRoute[Page.Main]}>
               <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
-            </a>
+            </Link>
           </div>
           {currentPage === Page.Login
             ? ''
