+++ src/components/private-route/private-route.tsx
@@ -1,0 +1,17 @@
+import {Navigate} from 'react-router-dom';
+import {AppRoute, Page, AuthorizationStatus} from '../../utils/const';
+
+type PrivateRouteProps = {
+  authorizationStatus: AuthorizationStatus;
+  children: JSX.Element;
+}
+
+export default function PrivateRoute(props: PrivateRouteProps): JSX.Element {
+  const {authorizationStatus, children} = props;
+
+  return (
+    authorizationStatus === AuthorizationStatus.Auth
+      ? children
+      : <Navigate to={AppRoute[Page.Login]} />
+  );
+}
